requirements.txt
python-dotenv
langchain
langgraph
langchain-google-genai
google-generativeai
langchain-community
tavily-python
trulens-eval
pydantic